Start adding navigability of pages.
Added a custom front-matter attribute called 'nav_position' that, if present,
causes a page to be included in the navigation for the site. In addition, the
value is used to sort the pages in the order they should appear.

Added a 'directions' page so that I could test it out.